cmake: cleanup CachingFileMgrTest linking
CachingFileMgrTest was being linked to way more libs than necessary,
which was causing issues with newer compilers (clang 11).

Solution is to only link to the libs related to the headers we're
directly including. This exposed some issues like Shared not being
directly linked to OSDependent and DataMgr not linked to Folly (if
enabled), even though they depend on them.