Move prepare_foreign_table_for_execution out from RelElgExecutor.
RelAlgExecutor performs actions specific to storage implementation
details. RelAlgExecutor should be abstracted from storage and therefore
this code is moved to DBHandler which is a better place for such code.

It will be probably better to introduce API in Catalog/DataMgr to
perform some preparation and finalization actions for queries.

Signed-off-by: ienkovich <ilya.enkovich@intel.com>